Split the tasks for moving and storing image info
I've changed the FolderSync settings to upload to the '_uploads' folder
on S3, and split the tasks to move the files and then separately
download the image information. Prior to this, I was only recording the
image information if I also moved the image.

I'm living with the duplication for now but plan to refactor later.