Add script that uses s3cmd to record image information
The s3 gem I'm using doesn't appear to understand folders in s3, so it
has to iterate over every object in s3 which is unsustainable.

Using the s3cmd, I can reduce the number of requests by first finding
the top-level date folders and then only iterating over the images in
those date folders that I haven't yet recorded in the database.