Update schedule.rb
The move-images and record-image-information tasks haven't been running
automatically for quite a while.

I think I broke this in the update on 1st Jan[1]. It would appear that
the `runner` command doesn't like running scripts in directories.
Having re-read the `whenever` README[2], it looks like `runner` is
designed to work with methods rather than filenames.

[1] 776b8afd5dde097422c3edf5db674c74d1240889
[2] https://github.com/javan/whenever