Display thumbnail images
I was previously relying on the browser to scale the images from their
original size of 2048x1536. This was only ever meant to be a temporary
measure and I'm now resizing the images in the process-images script so
I can use those instead. I'm currently running the
generate-other-image-sizes.rb script to resize all existing images so
there'll be missing images on the site for a while.